Add test use cases
I like to start out with tests and there's no better way than
starting out with the use cases. This helps me flush out any
missunderstanding or missing information for the problem at the earliest
possible time.

In this case, it was great that the use cases were given. At the moment
I am coding for the given happy path, later I will add use cases that
should not be palindromes in any base (given the input limit.)