Add test + specific solution for the decimal base
This commit is mainly so I can add the test, and give myself an easy
win and help me think about the problem by having passing tests for
the decimal base palindromes.

Moving forward, I will be working on code to represent the input in
different bases. Ultimately, since I'll need a solution that will check for
palindromes in any base, I'll try to come up with a generic solution.