Find palindromes' smallest bases
This took more thinking than I originally thought it would. I'm glad for
a wiki article on Positional_notation/Base_conversion that was of great
help.

In this commit my biggest win was solving the problem. My second biggest
win was introducing some OO features, which I think made most of
the code look cohesive.

The test suite was modified to fit the generic implementation. The happy
path is passing with all the given input. Almost 100% of the code has test coverage either
directly or indirectly.

Going forward next: will add test cases for input that should not be
palindromes, and add corresponding tests for each function not yet directly
tested.